requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
-e git+https://github.com/edx/analytics-python.git@1.2.11#egg=analytics-python==1.2.11
alabaster==0.7.12         # via sphinx
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
appdirs==1.4.3            # via zeep
asn1crypto==0.24.0        # via cryptography
astroid==1.4.9            # via pylint
babel==2.6.0              # via django-oscar, django-phonenumber-field, sphinx
backports.functools-lru-cache==1.5  # via pylint, soupsieve
beautifulsoup4==4.7.1     # via webtest
billiard==3.3.0.23        # via celery
cached-property==1.5.1    # via zeep
celery==3.1.26.post2      # via edx-ecommerce-worker
certifi==2018.11.29       # via requests
cffi==1.12.2              # via cryptography
chardet==3.0.4            # via requests
configparser==3.7.3       # via pylint
cookies==2.2.1            # via responses
coreapi==2.3.3
coreschema==0.0.4         # via coreapi
coverage==4.3.4
cryptography==2.6.1       # via pyopenssl
cssselect==1.0.3          # via premailer
cssutils==1.0.2           # via premailer
ddt==1.1.1
defusedxml==0.5.0         # via zeep
diff-cover==0.9.6
django-appconf==1.0.3     # via django-compressor
django-compressor==2.2
django-cors-headers==2.5.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.7
django-extra-views==0.10.0  # via django-oscar
django-filter==1.0.4
django-haystack==2.8.1    # via django-oscar
django-libsass==0.5
django-nose==1.4.4
django-oscar==1.5.4
django-phonenumber-field==1.3.0  # via django-oscar
django-rest-swagger==2.2.0
django-solo==1.1.3
django-tables2==1.16.0    # via django-oscar
django-threadlocals==0.10
django-treebeard==4.3     # via django-oscar
django-waffle==0.14.0
django-webtest==1.9.2
django-widget-tweaks==1.4.3  # via django-oscar
django==1.11.29
django_extensions==1.9.9
djangorestframework-csv==2.1.0
djangorestframework-jwt==1.11.0
djangorestframework==3.6.4
docutils==0.14            # via sphinx
drf-extensions==0.3.1
edx-auth-backends==1.2.1
edx-django-release-util==0.3.1
edx-django-sites-extensions==2.4.0
edx-django-utils==1.0.3
edx-drf-extensions==2.0.1
edx-ecommerce-worker==0.7.2
edx-i18n-tools==0.4.8
edx-opaque-keys==0.4.4
edx-rest-api-client==1.9.2
edx-sphinx-theme==1.0.2
enum34==1.1.6             # via cryptography
factory-boy==2.8.1
faker==1.0.2              # via factory-boy
freezegun==0.3.7
funcsigs==1.0.2           # via mock
future==0.17.1            # via pyjwkest
gevent==1.0.2
greenlet==0.4.15          # via gevent
httpretty==0.8.14
idna==2.8                 # via requests
imagesize==1.1.0          # via sphinx
inflect==2.1.0            # via jinja2-pluralize
ipaddress==1.0.22         # via cryptography, faker
isodate==0.6.0            # via zeep
isort==4.2.5
itypes==1.1.0             # via coreapi
jinja2-pluralize==0.3.0   # via diff-cover
jinja2==2.10              # via coreschema, diff-cover, jinja2-pluralize, sphinx
jsonfield==1.0.3
kombu==3.0.37             # via celery
lazy-object-proxy==1.3.1  # via astroid
libsass==0.9.2
lxml==3.8.0
markdown==2.6.9
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via pylint
mock-django==0.6.9
mock==1.3.0
ndg-httpsclient==0.5.1
newrelic==4.14.0.115      # via edx-django-utils
nose-ignore-docstring==0.2
nose==1.3.7               # via django-nose
oauthlib==3.0.1           # via requests-oauthlib, social-auth-core
openapi-codec==1.3.2      # via django-rest-swagger
path.py==7.2
paypalrestsdk==1.13.1
pbr==5.1.3                # via mock, stevedore
pep8==1.6.2
phonenumberslite==8.10.6  # via django-phonenumber-field
pillow==5.4.1             # via django-oscar
polib==1.1.0              # via edx-i18n-tools
premailer==2.9.2
psutil==1.2.1             # via edx-django-utils, edx-drf-extensions
ptvsd==3.0
purl==1.4                 # via django-oscar
pyasn1==0.4.5             # via ndg-httpsclient
pycountry==17.1.8
pycparser==2.19           # via cffi
pycryptodomex==3.7.3      # via pyjwkest
pygments==2.3.1
pyinotify==0.9.6
pyjwkest==1.3.2           # via edx-drf-extensions, social-auth-core
pyjwt==1.7.1              # via djangorestframework-jwt, edx-auth-backends, edx-rest-api-client, social-auth-core
pylint==1.6.4
pymongo==3.7.2            # via edx-opaque-keys
pyopenssl==19.0.0         # via ndg-httpsclient, paypalrestsdk
python-dateutil==2.8.0
python-openid==2.2.5      # via social-auth-core
pytz==2016.10
pyyaml==3.13              # via edx-django-release-util, edx-i18n-tools
rcssmin==1.0.6            # via django-compressor
requests-oauthlib==1.2.0  # via social-auth-core
requests-toolbelt==0.9.1  # via zeep
requests==2.21.0
responses==0.5.1
rest-condition==1.0.3     # via edx-drf-extensions
rjsmin==1.0.12            # via django-compressor
sailthru-client==2.2.3
selenium==3.4.3
semantic-version==2.6.0   # via edx-drf-extensions
simplejson==3.16.0        # via django-rest-swagger, sailthru-client
six==1.12.0
slumber==0.7.1            # via edx-rest-api-client
snowballstemmer==1.2.1    # via sphinx
social-auth-app-django==1.2.0  # via edx-auth-backends
social-auth-core[openidconnect]==1.7.0  # via edx-auth-backends, social-auth-app-django
sorl-thumbnail==12.5.0    # via django-oscar
soupsieve==1.8            # via beautifulsoup4
sphinx==1.5.3
sqlparse==0.2.4           # via django-debug-toolbar
stevedore==1.30.1         # via edx-opaque-keys
stripe==1.70.0
testfixtures==4.5.0
text-unidecode==1.2       # via faker
transifex-client==0.12.2
typing==3.6.6             # via django-extensions
unicodecsv==0.14.1
unidecode==0.4.21         # via django-oscar
uritemplate==3.0.0        # via coreapi
urllib3==1.24.1           # via requests, transifex-client
waitress==1.2.1           # via webtest
webob==1.8.5              # via webtest
webtest==2.0.33           # via django-webtest
wrapt==1.11.1             # via astroid
zeep==2.1.1
razorpay==1.2.0